pyproject.toml
[project]
name = "hitsz_manager"
version = "0.1.0"
description = "hitsz_manager"
readme = "README.md"
requires-python = ">=3.9, <4.0"
dependencies = [
    "nonebot2[fastapi]>=2.4.4",
    "nonebot-adapter-onebot>=2.4.6",
    "nonebot-plugin-status>=0.9.0",
    "nonebot-plugin-apscheduler>=0.5.0",
    "nonebot-plugin-alconna>=0.60.4",
    "httpx>=0.27.0",
    # --- 新增依赖 ---
    "GitPython>=3.1.0",           # 用于拉取 git 仓库
    "langchain>=0.1.0",           # RAG 框架
    "langchain-community>=0.0.10",
    "langchain-openai>=0.0.5",    # 调用兼容 OpenAI 接口的 LLM
    "langchain-huggingface>=0.0.1", # 调用 HuggingFace Embedding
    "chromadb>=0.4.22",           # 向量数据库
    "sentence-transformers>=2.3.0", # 本地 Embedding 模型支持
    "thefuzz>=0.19.0"             # 模糊匹配 (你之前已经有了，确保写上)
]

[tool.nonebot]
plugin_dirs = ["src/plugins"]
builtin_plugins = []

[tool.nonebot.adapters]
nonebot-adapter-onebot = [
    { name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11" }
]

[tool.nonebot.plugins]
nonebot-plugin-status = ["nonebot_plugin_status"]
nonebot-plugin-apscheduler = ["nonebot_plugin_apscheduler"]
nonebot-plugin-alconna = ["nonebot_plugin_alconna"]
course_manager = ["course_manager"]
pr_entry = ["pr_entry"]
